+++ src/DBFTPlugin/RecoveryMessage.cs
@@ -9,13 +9,13 @@ namespace Neo.Consensus
 {
     public partial class RecoveryMessage : ConsensusMessage
     {
-        public Dictionary<int, ChangeViewPayloadCompact> ChangeViewMessages;
+        public Dictionary<byte, ChangeViewPayloadCompact> ChangeViewMessages;
         public PrepareRequest PrepareRequestMessage;
         /// The PreparationHash in case the PrepareRequest hasn't been received yet.
         /// This can be null if the PrepareRequest information is present, since it can be derived in that case.
         public UInt256 PreparationHash;
-        public Dictionary<int, PreparationPayloadCompact> PreparationMessages;
-        public Dictionary<int, CommitPayloadCompact> CommitMessages;
+        public Dictionary<byte, PreparationPayloadCompact> PreparationMessages;
+        public Dictionary<byte, CommitPayloadCompact> CommitMessages;
 
         public override int Size => base.Size
             + /* ChangeViewMessages */ ChangeViewMessages?.Values.GetVarSize() ?? 0
@@ -41,8 +41,8 @@ namespace Neo.Consensus
                     PreparationHash = new UInt256(reader.ReadFixedBytes(preparationHashSize));
             }
 
-            PreparationMessages = reader.ReadSerializableArray<PreparationPayloadCompact>(byte.MaxValue).ToDictionary(p => (int)p.ValidatorIndex);
-            CommitMessages = reader.ReadSerializableArray<CommitPayloadCompact>(byte.MaxValue).ToDictionary(p => (int)p.ValidatorIndex);
+            PreparationMessages = reader.ReadSerializableArray<PreparationPayloadCompact>(byte.MaxValue).ToDictionary(p => p.ValidatorIndex);
+            CommitMessages = reader.ReadSerializableArray<CommitPayloadCompact>(byte.MaxValue).ToDictionary(p => p.ValidatorIndex);
         }
 
         public override bool Verify(ProtocolSettings protocolSettings)
